requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pycryptodome==3.23.0
Werkzeug==3.1.3